feat(issue-318-pr-c): retire geo_foundation from utilside (geo_io/entity/cam_core); add triangle + entity traits to geo_contracts
