+++ model/geo_algorithms/src/lib.rs
@@ -55,7 +55,7 @@ pub use octree::{Octree, OctreeNode, OctreeTolerance};
 // geo_foundationからの基本型の再エクスポート
 
 // NURBS型の再エクスポート（ViewModel層からのアクセス用）
-pub use geo_foundation::Scalar;
+pub use geo_contracts::Scalar;
 pub use geo_nurbs::adaptive_tessellation;
 pub use geo_nurbs::{NurbsCurve3D, NurbsSurface3D};
