+++ model/geo_contracts/src/entity/mod.rs
@@ -1,0 +1,4 @@
+//! geo_contracts entity module
+pub mod entity_traits;
+
+pub use entity_traits::{EntityDisplayProperties, EntityIdentity, LineEntity3DProperties};
