+++ model/geo_contracts/src/lib.rs
@@ -4,6 +4,7 @@
 //! `geo_primitives`, `geo_nurbs`, and higher-level algorithm crates.
 
 pub mod classification;
+pub mod entity;
 pub mod geometry;
 
 pub use analysis::abstract_types::{Angle, Scalar};
@@ -18,8 +19,11 @@ pub use geometry::core::{
     Vector2DMeasure, Vector2DProperties, Vector3DConstructor, Vector3DCore, Vector3DMeasure,
     Vector3DProperties,
 };
+pub use entity::{EntityDisplayProperties, EntityIdentity, LineEntity3DProperties};
 pub use geometry::core::{
     NurbsCurve2DConstructor, NurbsCurve2DCore, NurbsCurve2DMeasure, NurbsCurve2DProperties,
     NurbsCurve3DConstructor, NurbsCurve3DCore, NurbsCurve3DMeasure, NurbsCurve3DProperties,
     NurbsSurface3DConstructor, NurbsSurface3DCore, NurbsSurface3DMeasure, NurbsSurface3DProperties,
+    Triangle2DConstructor, Triangle2DCore, Triangle2DMeasure, Triangle2DProperties,
+    Triangle3DConstructor, Triangle3DCore, Triangle3DMeasure, Triangle3DProperties,
 };
