+++ model/geo_entity/src/geometric_entity.rs
@@ -1,5 +1,5 @@
 use crate::{Attributes, DisplayAttributes, EntityId, Metadata};
-use geo_foundation::{EntityDisplayProperties, EntityIdentity, LineEntity3DProperties, Scalar};
+use geo_contracts::{EntityDisplayProperties, EntityIdentity, LineEntity3DProperties, Scalar};
 
 #[derive(Debug, Clone)]
 pub struct GeometricEntity<T: Scalar, G> {
