+++ model/geo_foundation/src/entity/core/entity_traits.rs
@@ -1,23 +1,5 @@
-//! Entity Core Traits
+//! Entity Core Traits - geo_contracts への移設に伴う互換 re-export
 //!
-//! Foundation Pattern に基づく Entity 抽象化トレイト。
-//! ViewModel 層はこのトレイト経由で Entity を扱い、
-//! 具体実装（geo_entity など）への直接依存を避ける。
-
-use crate::Scalar;
-
-pub trait EntityIdentity {
-    type Id: Copy + Eq + std::hash::Hash;
-
-    fn entity_id(&self) -> Self::Id;
-}
-
-pub trait EntityDisplayProperties {
-    fn entity_visible(&self) -> bool;
-    fn entity_color(&self) -> [f32; 4];
-}
-
-pub trait LineEntity3DProperties<T: Scalar>: EntityIdentity + EntityDisplayProperties {
-    fn line_start(&self) -> (T, T, T);
-    fn line_end(&self) -> (T, T, T);
-}
+//! 正規参照先: `geo_contracts::entity::entity_traits`
+//! このファイルは後方互換のため `geo_contracts` からの再エクスポートとして機能する。
+pub use geo_contracts::entity::entity_traits::*;
