+++ model/geo_io/Cargo.toml
@@ -6,8 +6,8 @@ description = "File I/O operations for geometric data (STL, OBJ, PLY, etc.)"
 license = "MIT OR Apache-2.0"
 
 [dependencies]
+geo_contracts = { path = "../geo_contracts" }
 geo_primitives = { path = "../geo_primitives" }
-geo_foundation = { path = "../geo_foundation" }
 logging_foundation = { path = "../../foundation/logging_foundation" }
 tracing.workspace = true
