+++ model/geo_io/tests/stl_tests.rs
@@ -1,6 +1,6 @@
 //! STLローダーのテスト
 
-use geo_foundation::core::triangle_traits::Triangle3DProperties;
+use geo_contracts::Triangle3DProperties;
 use geo_io::stl;
 use geo_primitives::{Point3D, TriangleMesh3D};
 use std::io::Write;
